feat: support negative position decimal places

+++ core/execution/market.go
@@ -292,7 +292,7 @@ func NewMarket(
 		return nil, ErrEmptyMarketID
 	}
 
-	positionFactor := num.DecimalFromFloat(10).Pow(num.DecimalFromInt64(int64(mkt.PositionDecimalPlaces)))
+	positionFactor := num.DecimalFromFloat(10).Pow(num.DecimalFromInt64(mkt.PositionDecimalPlaces))
 
 	tradableInstrument, err := markets.NewTradableInstrument(ctx, log, mkt.TradableInstrument, oracleEngine)
 	if err != nil {
